requirements.txt
certifi==2020.6.20
cycler==0.11.0
dataclasses==0.8
importlib-resources==5.4.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
opencv-python==4.5.5.64
Pillow==8.4.0
pyparsing==3.0.7
python-dateutil==2.8.2
scipy==1.5.4
six==1.16.0
torch==1.10.1
torchvision==0.11.2
tqdm==4.63.0
typing_extensions==4.1.1
zipp==3.6.0
